requirements.txt
# Core ML
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0

# Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Transformers & Training
transformers>=4.35.0
peft>=0.6.0
sentence-transformers>=2.2.0

# Audio
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.0
pyaudio>=0.2.13

# Utilities
tqdm>=4.66.0
